DUOS-760 pulled out styles and placed in Theme file, pulled out modal into its own file, removed uneeded variables from chair console, comment cleanup
