+++ src/libs/theme.js
@@ -59,5 +59,166 @@ export const Theme = {
     fontSize: '16px',
     padding: '10px',
     fontWeight: '500'
+  },
+};
+
+export const Styles = {
+  PAGE: {
+    width: "90%",
+    margin: "0 auto"
+  },
+  TITLE: {
+    fontFamily: "Montserrat",
+    fontWeight: Theme.font.weight.semibold,
+    fontSize: Theme.font.size.title,
+  },
+  SMALL: {
+    fontFamily: 'Montserrat',
+    fontWeight: Theme.font.weight.regular,
+    fontSize: Theme.font.size.small
+  },
+  HEADER_IMG: {
+    width: '60px',
+    height: '60px',
+  },
+  HEADER_CONTAINER: {
+    display: 'flex',
+    flexDirection: "column"
+  },
+  ICON_CONTAINER: {
+    flexBasis: '76px',
+    height: '60px',
+    paddingRight: '16px'
+  },
+  RIGHT_HEADER_SECTION: {
+    display: 'flex',
+    alignItems: 'flex-end'
+  },
+  LEFT_HEADER_SECTION: {
+    display: 'flex',
+    flexDirection: 'row',
+    paddingTop: "3rem"
+  },
+  TABLE: {
+    CONTAINER: {
+      margin: "3rem auto"
+    },
+    HEADER_ROW: {
+      fontFamily: "Montserrat",
+      fontSize: "14px",
+      color: "#00243C",
+      fontWeight: Theme.font.weight.medium,
+      backgroundColor: "#f3f6f7",
+      display: "flex",
+      justifyContent: "center",
+      height: "51px"
+    },
+    RECORD_ROW: {
+      fontFamily: 'Montserrat',
+      fontWeight: Theme.font.weight.regular,
+      fontSize: "14px",
+      display: "flex",
+      justifyContent: "center",
+      height: "48px",
+    },
+    RECORD_TEXT: {
+      color: "#00243C"
+    },
+    DAR_TEXT_HOVER: {
+      cursor: 'pointer',
+      color: '#0cabc5d9'
+    },
+    DATA_REQUEST_TEXT: {
+      color: "#00609F",
+      fontWeight: Theme.font.weight.semibold
+    },
+    //NOTE: play around with the cell measurements
+    TITLE_CELL: {
+      width: "18%",
+      display: "flex",
+      justifyContent: "left",
+      alignItems: "center",
+      margin: "0 2%"
+    },
+    DATA_ID_CELL: {
+      width: "10%",
+      margin: "0 2%",
+      display: "flex",
+      justifyContent: "left",
+      alignItems: "center",
+    },
+    SUBMISSION_DATE_CELL: {
+      width: "10%",
+      margin: "0 2%",
+      display: "flex",
+      justifyContent: "left",
+      alignItems: "center",
+    },
+    DAC_CELL: {
+      width: "10%",
+      margin: "0 2%",
+      display: "flex",
+      justifyContent: "left",
+      alignItems: "center",
+    },
+    ELECTION_STATUS_CELL: {
+      width: "10%",
+      margin: "0 2%",
+      display: "flex",
+      justifyContent: "left",
+      alignItems: "center",
+    },
+    ELECTION_ACTIONS_CELL: {
+      width: "23%",
+      margin: "0 2%",
+      display: "flex",
+      justifyContent: "left",
+      alignItems: "center",
+    }
+  },
+  MODAL: {
+    CONTENT: {
+      height: "auto",
+      top: "20%",
+      bottom: "20%",
+      left: "20%",
+      right: "20%",
+      fontFamily: 'Montserrat',
+      padding: "2%"
+    },
+    DAR_SUBHEADER: {
+      display: 'flex',
+      fontFamily: 'Montserrat',
+      fontSize: '16px',
+      fontWeight: Theme.font.weight.semibold,
+      justifyContent: 'left',
+      color: "#777"
+    },
+    TITLE_HEADER: {
+      display: 'flex',
+      fontFamily: 'Montserrat',
+      fontSize: Theme.font.size.title,
+      fontWeight: Theme.font.weight.regular,
+      justifyContent: 'left',
+      marginBottom: '4%'
+    },
+    DAR_DETAIL_ROW: {
+      padding: '0 3%',
+      margin: '2% 0',
+      display: 'flex',
+      justifyContent: 'space-between'
+    },
+    DAR_LABEL: {
+      color: "#777777",
+      fontSize: '14px',
+      fontWeight: Theme.font.weight.semibold,
+      width: "25%",
+      textAlign: 'right'
+    },
+    DAR_DETAIL: {
+      fontSize: '16px',
+      fontWeight: Theme.font.weight.medium,
+      width: "70%"
+    }
   }
 };
